+++ src/app/browsers/page.tsx
@@ -1,4 +1,4 @@
-import Browsers from '@/src/components/features/browsers/browsers';
+import Browsers from '@/components/features/browsers/browsers';
 
 export default function BrowsersPage() {
 	return <Browsers />;
