+++ src/app/editors/page.tsx
@@ -1,4 +1,4 @@
-import Editors from '@/src/components/features/editors/Editors';
+import Editors from '@/components/features/editors/Editors';
 
 export default function EditorPage() {
 	return <Editors />;
