+++ src/components/core/Banner.tsx
@@ -1,3 +1,4 @@
+import Link from 'next/link';
 import React from 'react';
 
 type BannerPropsTypes = {
@@ -18,19 +19,19 @@ function Banner({ titles, text }: BannerPropsTypes) {
 					<p className='mt-4 sm:text-l/relaxed'>{text}</p>
 
 					<div className='mt-8 flex flex-wrap justify-center gap-4'>
-						<a
+						{/* <Link
 							className='block w-full rounded bg-red-600 px-12 py-3 text-sm font-medium text-white shadow hover:bg-red-700 focus:outline-none focus:ring active:bg-red-500 sm:w-auto'
 							href='/get-started'
 						>
 							Get Started
-						</a>
+						</Link> */}
 
-						<a
+						{/* <a
 							className='block w-full rounded px-12 py-3 text-sm font-medium text-red-600 shadow hover:text-red-700 focus:outline-none focus:ring active:text-red-500 sm:w-auto'
 							href='/about'
 						>
 							Learn More
-						</a>
+						</a> */}
 					</div>
 				</div>
 			</div>
