+++ src/components/features/editors/Editors.tsx
@@ -1,39 +1,44 @@
 'use client';
-
 import EntityExtendMore from '../../common/EntityExtendMore/EntityExtendMore';
-import axios from 'axios';
-import { useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import Skeleton from '../../common/Skeleton/Skeleton';
 import Anchor from '../../core/Anchor';
-import { Editor } from './types';
+import { EditorTypes } from './types';
+import { API_ENDPOINTS } from '@/config/API.config';
+
+// SERVICES
+import { HttpService } from '@/services';
 
 function Editors() {
-	const [editors, setEditors] = useState([]);
+	const http = new HttpService();
+
+	const [editors, setEditors] = useState<EditorTypes[]>([]);
 	const [loading, setLoading] = useState(false);
 
-	useEffect(() => {
-		async function fetchEditorsData() {
-			try {
-				setLoading(true);
-				const { data } = await axios.get('/api/editors');
-				console.log({ data });
-				setEditors(data);
-				setLoading(false);
-			} catch (e) {
-				console.log(e);
-				setLoading(false);
-			}
-		}
+	const fetchEditorsData = useCallback(async () => {
+		try {
+			setLoading(true);
+			const data = await http.service().get<any[]>(API_ENDPOINTS.editors);
 
-		fetchEditorsData();
+			// console.log(data);
+			setEditors(data);
+			setLoading(false);
+		} catch (e) {
+			console.log(e);
+			setLoading(false);
+		}
 	}, []);
+
+	useEffect(() => {
+		fetchEditorsData();
+	}, [fetchEditorsData]);
 
 	return (
 		<section className='mt-10 p-10'>
 			<h3 className='text-2xl mb-3 font-bold leading-7 text-gray-900'>Editors</h3>
 			{loading ? <Skeleton type='image' /> : null}
-			{editors?.map(({ title, link, extensions, avatar }: Editor, index) => (
-				<div className='capitlise p-2' key={index}>
+			{editors?.map(({ title, link, extensions, avatar }: EditorTypes, index) => (
+				<div className='capitlise p-2  bg-slate-200 my-4' key={index}>
 					<div className='flex items-center justify-between gap-2 mb-5'>
 						<div className='flex items-center gap-4'>
 							<div className='shrink-0 flex items-center justify-center h-20 w-20 rounded bg-white border border-slate-200'>
@@ -53,23 +58,6 @@ function Editors() {
 									text: '',
 									list: extensions,
 								},
-								// {
-								// 	icon: 'anything',
-								// 	title: 'Flows',
-								// 	text: '',
-								// 	list: [
-								// 		{
-								// 			icon: 'anything',
-								// 			title: 'Analytics',
-								// 			text: 'Everything counts. And everything that counts shapes us.',
-								// 		},
-								// 		{
-								// 			icon: 'anything',
-								// 			title: 'Transform',
-								// 			text: 'Everything counts. And everything that counts shapes us.',
-								// 		},
-								// 	],
-								// },
 							]}
 						/>
 					) : null}
